% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{run_colorlab}
\alias{run_colorlab}
\alias{launch_palette_tutorial}
\title{Launch the ColorLab Tutorial in RStudio}
\usage{
run_colorlab(...)

launch_palette_tutorial(...)
}
\arguments{
\item{...}{Additional arguments forwarded to \code{learnr::run_tutorial()}.}
}
\value{
Runs the learnr interactive tutorial application.
}
\description{
Launches the shiny_prerendered learnr tutorial optimized for the RStudio
Tutorial Pane. Once launched, it appears under the 'Tutorial' tab in RStudio
or opens in a standalone web browser window.
}
\examples{
\dontrun{
ColorLab::run_colorlab()
}
}
